requirements.txt
django~=2.0
djangorestframework~=3.3
django-rest-auth~=0.7
python-social-auth[django]~=0.3
python3-openid==3.0.10
IGitt==0.4.1.dev20180608090851
django-cors-headers~=1.3.1
celery~=4.1.0
gunicorn~=19.7.1
gevent~=1.2.2
psycopg2==2.7.3
bugspots3==0.3.0.dev20171215101805
attrs~=17.4.0
django-pglocks~=1.0.2
raven~=6.4.0
coreapi~=2.3.3
